Use cobra command for argument handling
While this doesn't add any options or arguments, it does set up the
command to recognize them. Specifically, that it shouldn't receive
any arguments or options.

This results in an error message when a user executes the command with
additional arguments or options, so that usage like
`git lzc --version` won't result in the executable surprisingly
generating a commit.

Fixes #10